requirements.txt
requests
pyyaml
nose
